# swagger/song-library.yaml
openapi: 3.0.3
info:
  title: Music library
  description: Title
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: local server
paths:
  /songs:
    get:
      summary: Get songs from library with filtering and pagination
      parameters:
        - name: group
          in: query
          schema:
            type: string
          description: Filter by group name
        - name: song
          in: query
          schema:
            type: string
          description: Filter by songs title
        - name: date
          in: query
          schema:
            type: string
          description: Filter by songs release date
        - name: page
          in: query
          required: true
          schema:
            type: integer
          description: Page number for pagination
        - name: limit
          in: query
          required: true
          schema:
            type: integer
          description: Number of items per page
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  songs:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: The name of the song
                        group:
                          type: string
                          description: The group of the song
                        songDetail:
                          $ref: '#/components/schemas/SongDetail'
                  page:
                    type: integer
                    description: Page number for pagination
                  limit:
                    type: integer
                    description: Number of items per page
                  items:
                    type: integer
                    description: Number of returned items
        '204':
          description: No data. Songs not found
        '400':
          description: Bad request
        '500':
          description: Internal server error
    post:
      summary: Add a new songs to the library
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group:
                  type: string
                song:
                  type: string
              required:
                - group
                - song
      responses:
        '201':
          description: Song added successfully
        '400':
          description: Bad request
        '500':
          description: Internal server error
    put:
      summary: Update existing songs data
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group:
                  type: string
                song:
                  type: string
                songDetail:
                  $ref: '#/components/schemas/SongDetail'
              required:
                - group
                - song
                - songDetail
      responses:
        '200':
          description: Song updated successfully
        '400':
          description: Bad request
        '404':
          description: Song not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a songs from the library
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group:
                  type: string
                  description: Group of the song
                song:
                  type: string
                  description: Title of the song
              required:
                - group
                - song
      responses:
        '200':
          description: Song deleted successfully
        '204':
          description: Song to delete not found
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /songs/text:
    get:
      summary: Get lyrics of a specific songs with pagination
      parameters:
        - name: group
          in: query
          required: true
          schema:
            type: string
          description: Group of the songs
        - name: song
          in: query
          required: true
          schema:
            type: string
          description: Title of the songs
        - name: page
          in: query
          required: true
          schema:
            type: integer
          description: Page number for lyrics pagination
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: The name of the song
                  group:
                    type: string
                    description: The group of the song
                  text:
                    type: string
                    description: The verse of the song
                  page:
                    type: integer
                    description: Verse number of the song
        '204':
          description: No data found
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /info:
    get:
      parameters:
        - name: group
          in: query
          required: true
          schema:
            type: string
        - name: song
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongDetail'
        '204':
          description: No data found
        '400':
          description: Bad request
        '500':
          description: Internal server error
components:
  schemas:
    SongDetail:
      required:
        - releaseDate
        - text
        - link
      type: object
      properties:
        releaseDate:
          type: string
          example: 16.07.2006
        text:
          type: string
          example: Ooh baby, don't you know I suffer?\nOoh baby, can you hear me moan?\nYou caught me under false pretenses\nHow long before you let me go?\n\nOoh\nYou set my soul alight\nOoh\nYou set my soul alight
        link:
          type: string
          example: https://www.youtube.com/watch?v=Xsp3_a-PMTw